Breaks out non-setup functions into own file for test-ability
